requirements.txt
fastapi==0.75.0
python-dotenv==0.19.2
SQLAlchemy==1.4.32